requirements.txt
fastapi
pydantic
uvicorn
python-multipart
torch
torchvision
opencv-python
python-doctr